Re-organize models and services
